Add mysql service to docker compose
This is lightweight enough that it's worth adding now, just to make it easier for when I need to restore a mattermost mysql db backup to examine manually. Eventually there should be mmdata.py code to extract information from that database as well.

make pull-images will get the requisite mysql image.